+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## [0.0.1]
 
 - Initial release
+
+## [0.0.2]
+
+- Add `refresh()` function for `DocumentController`. 
+- Set a non-empty return value for `getScopeAt()` function.